extract google login to wrapper component
