+++ src/components/shared/GoogleLoginWrapper/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "GoogleLoginWrapper.jsx"
+}
